QNX: Add QQnxAudioRecorder class
This class will be used for capturing and encoding audio for an
audio-only QMediaSession session.

On QNX, camera audio is directly encoded by the camera/sensors API,
whereas audio-only encoding is done by mm-renderer.

QQnxAudioRecorder is a wrapper around mm-renderer based on
QQnxMediaPlayer (but with reduced functionality and reversed flow).

Change-Id: I23c11b91da720ba272a2fa04b58f5b53f264a62f
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>