Fix Camera Preview on declarative-camera
The example declarative camera showed a black screen in the camera
preview until a touch on the screen occurred. It was introduced by
36f110e6c97e2c98ded86945825aa563fd93f2e0 patch. This patch fixes
this issue.
While working on the issue, the fix uncovered another issue regarding
setting a camera format not correctly initialized. That is also fixed
on this patch.

Task-number: QTBUG-102029
Change-Id: I27adb97be2c2ace5f33818256e5b47ecf9e7210f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
(cherry picked from commit 540ff1496bcc9ef4019eb34978715b79ef0615a4)