QNX: Encapsulate mmrenderer code in QQnxMediaPlayer
Remove mmrenderer specific code from QQnxVideoSink. Video frames can
come from a variety of sources, not being limited to mm-renderer (e.g.
camera). It's important to keep QQnxVideoSink agnostic in such cases.

QQnxMediaPlayer already features most of the mm-renderer glue code. It
is therefore only logical that the remainder of the code, which is
solely used for media player, to be moved there.

Change-Id: I963f1c85c7c944ab9782185ff758bdfab984c86b
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>