Android: fix source device issue
Changed the QDataStream.write to the File.Write,
because a QDataStream.write add an extra 4 -bytes to the file.

For this reason the AMediaExtractor_setDataSourceFd will fail.

Also changed the tempPath to the QDir.tempPath(),
because a file creation fails with a tempPath;

Task-number: QTBUG-100079
Change-Id: Ieff2edec4f5e1b52fb55f6b79fc6944fda229bf7
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 3f880a5c66b5c9e005cff8eee296ef8b8b039563)