Add support for rooms to the spatial audio engine
The new QSpatialAudioRoom class allows defining a room.
If the listener is inside the room, the properties of that
room will get enabled, giving us first order reflections
and reverb.

Multiple rooms can be defined and the engine will automatically
update the room to be used when the listener position changes.
If the listener is inside multiple rooms, the room with the
smallest volume will get chosen.

Updated the spatialaudio example to allow modifying some
of the room properties for testing.

Change-Id: I94ffe0a6b7e570c4ea77898e021d7fe042fb9b52
Reviewed-by: Lars Knoll <lars.knoll@qt.io>