Update commercial license headers
Updated header.COMM to the files in tqtc-qtmultimedia.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4951
Change-Id: I01b40ad899401d0d0752603278182680ba262573
Reviewed-by: Lars Knoll <lars.knoll@qt.io>