tst_qmediaplayerbackend: don't check signal count
Do not set the signal count on the signal spies for the position and
state fields. When the media is done playing before the spies are
cleared, it is reasonable that the backends could be already on the
stopped state and on a fixed position (e.g. 0). Resetting the media
source shouldn't necessarily trigger a state change for these fields -
it in fact makes sense that the state remains untouched (e.g. 'Stopped',
or position == '0').

Checking the final state only (which already happens)is a better check
than relying on signal spy counts for this reason.

Change-Id: I49f720e640fa3733154e6ad19272f049b99994a1
Reviewed-by: Lars Knoll <lars.knoll@qt.io>