Windows: Fix audio buffer sizes
Audio buffer sizes were being calculated in a way that could result
in buffer sizes that were not a multiple of the audio frame size.
For instance, a buffer with an odd number of bytes could be used for
holding 16-bit audio data, which would result in the audio data
contained on half of buffers received by applications being shifted
by half a sample, causing applications to misinterpret that audio data.

Pick-to: 6.3 6.2
Fixes: QTBUG-102843
Change-Id: I252384cb99fffca5e068b0b14e4be9b0469962ec
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>