Fix compilation of resonance audio on our supported platforms
Disable neon optimizations on QNX. I can't get
those to compile on the platform. Looks like a bug
in the arm_neon.h headers on QNX.

Define M_PI and M_SQRT2 for pffft.c if they aren't defined.

Fix the arm processor detection to be aligned with what we
have in qprocessordetection.h

Change-Id: I6f6ee2190ee6dd9c3a08d86961c39b47dd38f5cd
Reviewed-by: Lars Knoll <lars.knoll@qt.io>