Fix chunk sizing in QPulseAudioSink
QPulseAudioSink:write() takes a length parameter which is set to the
qMin of itself and the return value of pa_stream_writable_size().

QPulseAudioSink:write() takes a length parameter which is set to the
qMin of itself and the return value of pa_stream_writable_size().

This method returns the number of bytes requested by the server,
but not yet written. Unless data is buffered, the value is 0.
That leads to length being set to 0 and no data being sent.

In PulseOutputPrivate::writeData(), a loop was implemented calling
QPulseAudioSink::write() up to 10 times for each data chunk.
That forced data buffering and subsequently the return value of
pa_stream_writable_size() to be > 0.

This patch replaces the query of pa_stream_writable_size() with the
nbytes value modified by pa_stream_begin_write(). It represents the
maximum amount of bytes the pulseaudio server is ready to accept.

The patch also removes the loop calling write up to 10 times as it
becomes unnecessary.

Fixes: QTBUG-60575
Pick-to: 6.3 6.2 5.15
Change-Id: I316a5ac610c8540ab8f7fff4a3b0b1ee3d7e7ad5
Reviewed-by: Lars Knoll <lars.knoll@qt.io>