QtMultimedia: replace qSwap with std::swap/member-swap where possible
qSwap() is a monster that looks for ADL overloads of swap() and also
detects the noexcept of the wrapped swap() function, so it should only
be used when the argument type is unknown. In the vast majority of
cases, the type is known to be efficiently std::swap()able or to have
a member-swap. Call either of these.

Task-number: QTBUG-97601
Change-Id: Ida8402522bc980e871455d8fb21dddcf5b2dd384
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 583b991120596a5fbfdfd726ec228c23e819107e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>