Android: fix media devices race condition
There was a race condition regarding the QtAudioDeviceManager and
qandroidmediadevices. QtAudioDeviceManager registers a receiver for
audio devices that calls native functions when triggered.
Sometimes, depending on devices and the android system, it can call the
native functions before the functions are registered on
qandroidmediadevices.
Changed the register of QtAudioDeviceManager receiver to be done after
the register of native functions and only if succeeds.

Change-Id: I92aff9e9cbef13b209e1c459e308a447480293e6
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>