spatial audio: Use centimeters by default
Qt Quick 3D uses a centimeters by default for positions and
distances. Align our API with that choice.

Add a distanceScale property to enable the user to choose a
different scale.

Store all our data internally in meters, to be aligned with the
data expected by resonance audio and to make sure we don't
mess up the existing configuration when the distance scale changes.

Change-Id: Id06cb49bd152bfa7920c096c662ff103041261e0
Reviewed-by: Rafael Roquetto <rafael.roquetto@qt.io>