Enable HW encoding on Windows for ffmpeg
FFmpeg HW encoding uses COM objects in the COINIT_MULTITHREADED
concurrency mode. However, the main thread configures access to COM in
the COINIT_APARTMENTTHREADED mode with the call to OleInitialize, which
is not compatible with FFmpeg. As the concurrency model is defined per
thread, we can open the FFmpeg encoder in the encoder thread and let it
configure its desired mode.

Change-Id: I5c868f95c50228394ba87a9547083576cbdade4f
Reviewed-by: Lars Knoll <lars.knoll@qt.io>