PulseAudio: fix sink transition to idle
When the client writing to the sink has no more data, then the sink
will keep rendering the data already written until it runs into buffer
underflow. It will then call the relevant callback.

At that point, check if the source device has reached the end. If it
has, then wait for the stream to be drained before signalling the state
change to idle.

Pick-to: 6.3
Change-Id: If7d626ff2f31460abf61971e27f992db2af534a0
Reviewed-by: Lars Knoll <lars.knoll@qt.io>