Fix state transition in QPulseAudioSink::resume()
Calling resume() resulted in a state transition depending on the audio
sink's mode: Transition to ActiveState in pull mode and to IdleState in
push mode - even when there was audio playback remaining.
This patch always transitions the state to ActiveState.
State transitions to IdleState anyway, when the playback has finished.

Change-Id: Idcd0be16929ea430cddc99a9a0284ea75e9b335b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>