Fix state transition in QGStreamerAudioSink::resume()
Calling resume() resulted in a state transition depending on the audio
sink's mode: Transition to ActiveState in pull mode and to IdleState in
push mode - even when there was audio playback remaining.
This patch always transitions the state to ActiveState.
State transitions to IdleState anyway, when the playback has finished.

Change-Id: I34146eeab81f5d6629ac90c5153abc3e4e6e77c4
Reviewed-by: Lars Knoll <lars.knoll@qt.io>