Spatial Audio: Implement support for room reverb in surround mode
Resonance audio doesn't calculate any reverb when we ask for
ambisonic output. The reason is that reverb is purely calculated
using mono sources and a stereo output and is directly
connected to the stereo mixer at the end of the pipeline.

To fix this, gain access to private parts of resonance audio by
compiling it with -Dprivate=public. That's ok, as we compile it purely
as an internal library anyway.

After that retrieve the audio data for the reverb component from the
graph manager and add it at the end of the processing pipeline (after
ambisonic decoding) to the different surround channels.

Change-Id: Iaeeec8df7a963858b6ec0399948a55f89136964e
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>