Patch resonance audio to give access to reverb data
When retrieving the ambisonic sound field from resonance audio,
that data does contain room reflections, but not reverb for
the room.

The reason is that reverb is directionally independent and
resonance audio only calculates it in stereo.

Work around this, by adding a small getter that allows us to
retrieve the audio buffer that contains the reverb data.

Change-Id: I6dd177293d38713ce29c40483dfc8d2409828fce
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>