Android: Fix for the permissions issue on the tst_QAudioSource test
When testing for the first time with an emulator with Android version <= 9 after test installation,
the test passes because we do not have record permissions from user, so we use the default SampleRates list for QAudioDevice input.
We don’t even check to see if these sample rates are supported in opensles.
After the fix, the QOpenSLESEngine::supportedSampleRates() function works the same way on the first and second runs of the test and so on.

Fixes: QTBUG-99095
Pick-to: 6.2 6.3
Change-Id: I0b61232b5cb38c314913fa60c8bf2fd374b0468f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>