Add basic screen capturing support using QScreen::grabWindow
This patch adds a basic screen capturing support. It adds a basic API
to enable capturing of a screen pointed to by QScreen, QWindow,
or a window pointed to by WId on QScreen.

The patch implementes a basic QScreen capturing using
the QScreen::grabWindow function running continuously in a separate
thread and fed into FFmpeg encoder. It does not support window
capturing yet.

Initial testing indicates that the grabWindow function has poor
performanc; thus, we need to gradually replace it with platform
depenedent code optimized for screen capturing.

Change-Id: I0d2ec1a2f211edf287f7d21e73eead26202aacdc
Reviewed-by: André de la Rocha <andre.rocha@qt.io>