Android: Fix for UnderrunError on QAndroidAudioSink
After fix we set a state of QAndroidAudioSink to the QAudio::IdleState,
immediately when m_audioSource->atEnd() is true and buffer size is > 0.
Because otherwise it is called when buffer size is 0 and m_audioSource->atEnd() is true.
and it leads to that we also call a setError(QAudio::UnderrunError)
in QAndroidAudioSink::onEOSEvent() function.

The patch will not affect the tests.

Task-number: QTBUG-97492
Pick-to: 6.2 6.3 6.4
Change-Id: Id48469141ab4aab266b775aea674bc6e82a53d33
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>