QtMultimedia - fix a crash on iOS
While the key for type format is documented as either corresponding to
an NSNumber or an array of numbers, the AVFoundation crashes when calling
(non-existing) selector - __NSSingleItemArrayI does not have the
selector 'intValue' (which NSNumber does have). Also, warnings
from AVFoundation suggest that OpenGL and Metal Compatibility
keys are not supported (on iOS, but no warnings on macOS).

Fixes: QTBUG-103725
Change-Id: I078b43f5d7d1f83e7d3910894efd25f8122b4e8f
Reviewed-by: Lars Knoll <lars.knoll@qt.io>