Provide QRhi device when mapping QVideoFrame textures onto that QRhi
The textures of QVideoFrames are consumed by a QRhi device that calls
textureHandle, or texture on the frame. However, this QRhi device
is not explicitly signalled to QVideoFrame when preparing the texutre.
Instead, QVideoFrame uses QRhi that it fetched from QMediaPlayer or
QMediaCaptureSession context related to QVideoSink. This should work
for most cases.
This patch explicitly states to which QRhi device QVideoFrame textures
should be mapped. This can be later used to execute or validate
the mapping, which is not a part of this patch.

Change-Id: I4e0301f4ebbc68ec6d482cf52dc40f10d59c7dce
Reviewed-by: André de la Rocha <andre.rocha@qt.io>