EVR: clear the event queue before deallocation
In the windows EVR implementation, EVRCustomPresenter instances are
ref-counted and freed by invoking the dtor directly. Some events may
still be in the queue however, and in some cases seem to be delivered
to the instance after the dtor is done and has freed members, resulting
in a crash.
Using deleteLater() ensures the event queue is cleared before the dtor
is executed.

Change-Id: Ibbbb18ac97a67a58e4b77d8c57f97408d07b8f8a
Pick-to: 6.3 6.4
Reviewed-by: André de la Rocha <andre.rocha@qt.io>