Android: Fix for changing Audio sink
The OpenSL ES API does not provide direct audio output IO-devices as a
sink for media objects. That's why setting audio output in
QAndroidAudioSink is done with JNI calls by AudioManager.

The disadvantage of this solution is that the audio output is set for the
entire Android device (not just for the current sink). But this is still
a better solution than not working audio output change at all.

Pick-to: 6.4 6.3 6.2
Task-number: QTBUG-99022
Change-Id: Id9a3d7a1abf2497b8e94363c93f888d09db9edb4
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>