Avoid queueing video frames
When sending QVideoFrame with a queued connection, a large number
of frames can queue up waiting for execution of the signal handler.
This leads to significat resource consumption that can in turn cause
crash. For instance, when moving the window of the media player example
for 2sec with FFmpeg backend on Windows HW codec crashed.

This patch handles videFrameChanged signal carrying a video frame in
the lambda expression. The expression is executed on the thread emitting
the frame; thus, we can immediately update the pending frame (m_frame)
and avoid queueing frames on the queued connection connected to
the _q_newFrame slot.

Change-Id: I87c71afa404d7974cd4585906f53051a65b08f4b
Reviewed-by: Lars Knoll <lars.knoll@gmail.com>
(cherry picked from commit 5fc1e9179426fefe2ad9444961d7e66baa159dd3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>