Android: Fix framerate drop when QML animation is over video/camera
The reason for framerate drop was the invoke method that forces to
readback a new frame from the surface. Removing that method may lead
for the video output to be become stuck when the render thread becomes
blocked, since because the object needs to be moved to OpenGL thread to
be able to create OpenGLContext. To solve this, it now verifies if the
thread needs to create OpenGLContext and moves to OpenGL thread but will
move back to the original thread after the OpenGLContext are created.

Fixes: QTBUG-104133
Change-Id: I38517e8c1e9db2f4098aaaaa5a0598a406f04281
Reviewed-by: Lars Knoll <lars.knoll@gmail.com>
(cherry picked from commit e8248b7a953ca2284f4fa489dfc31579bb92f66e)
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>