Export textures in a bundle as QVideoFrameTextures to requesting QRhi
This patch adds the QVideoFrameTextures class that holds textures
created for a given QRhi device and takes responsibiliy of cleaning
up after disposing of them. Exporting texture handlers as it was
implemented does not work as these are often not created for QRhi that
will use them.

For example, when calling QVideoFrame::toImage() and using HW
accelerated conversion, we have to generate a new set of texture handles
that are compatible with QRhi used for conversion. In parallel
a frame may be displayed using other QRhi that requires other set
of handles that are created specifically for it.

Change-Id: I73f99ca2c17207f403409228b933dd76671451c7
Task-number: QTBUG-105169
Reviewed-by: André de la Rocha <andre.rocha@qt.io>