Run Android video output from a dedicated thread and shared OpenGL ctx
This patch restructures QAndroidVideoOutput threading. It avoids
scheduling frame readbacks on a rendering threads. Instead, it creates
a dedicated thread producing video frames and shares that thread OpenGL
contexts with all threads that use/consume the texture.

Additionally, this patch corrects the behavior of QVideoFrame where
all QVideoFrames created by QAndroidVideoOutput carried the same
texture. This was casued by the limitation of AndroidSurfaceTexture
that makes available at any point in time only one frame as
an GL_TEXTURE_EXTERNAL_OES texture that cannot be copied.
This patch renders this frame into a normal texture, which is then
attached to QVideoFrame.

Change-Id: Ib0782f60d5a37dc8ca9c897143dd845782c77403
Task-number: QTBUG-105169
Reviewed-by: Lars Knoll <lars.knoll@gmail.com>