WMF plugin: fix race condition on acivating/stopping media session
In the case when the media session stops immediately after starting,
we can get into a situation where it is still active (m_active == true).
This is because QWindowsMediaDeviceSession::handleStreamingStarted() is
called through the event loop because the corresponding signal
is emitted from another thread. To fix the problem we need to handle
this signal only in "activating" state and ignore otherwise.

Pick-to: 6.4
Change-Id: Ifb40158861885bbdd9cc847efefd6de1c442601e
Reviewed-by: André de la Rocha <andre.rocha@qt.io>