WMF plugin: fix race condition on acivating/stopping media session
In the case when the media session stops immediately after starting,
we can get into a situation where it is still active (m_active == true).
This is because QWindowsMediaDeviceSession::handleStreamingStarted() is
called through the event loop because the corresponding signal
is emitted from another thread. To fix the problem we need to handle
this signal only in "activating" state and ignore otherwise.

Change-Id: Ifb40158861885bbdd9cc847efefd6de1c442601e
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
(cherry picked from commit 35c0beaeea49a53574bfb3a9f998c59a34d9c161)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>