Validate that gstreamer elements necessary for operation are available
This patch validates that all elements that are necessary to run
the requested functionality are available. Instead of running
constructors directly, the patch uses a static create() function for
each class. This function first tries to instantiate all necessary
GStreamer plugins and then pass them to the class constructor.
If some element is missing the function returns an error message
describing the element.

Pick-to: 6.4
Fixes: QTBUG-106257
Change-Id: I2b9db3f509e9ad5b760fcc251e634a01c687b790
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>