AVFMediaPlayer: Avoid possible operation on dangling pointer
When submitting an async block to the dispatch queue in
AVFMediaPlayerObserver we are using a  __block variable
('blockSelf') to reference 'self' in order to not have a
strong reference to it inside the block, because otherwise
we could have a retain loop and a memory leak.

However, we risk that in a scenario where the observer gets
deleted, 'blockSelf' becomes a dangling pointer. The
'if (session)' check should protect against an operation
on the dangling pointer, because if the session is still
alive then the observer must be too (since the session
owns the observer and the observer is only released in
the session's destructor). Yet, it looks like something
goes wrong and we do end up in this scenario - the
method gets called on the 'blockSelf' (dangling) pointer,
even when the observer gets destroyed. This may be
related to the fact that the loadValuesAsynchronously
callback gets executed on a background thread, and
the callback block is capturing the 'session' QPointer,
copying it on that thread, possibly leading to a race
condition.

To avoid this, let's not rely on a check on the
QPointer at all. Instead explicitly retain the blockSelf
pointer to make sure that we don't end up with a
dangling pointer, and explicily release it after the
block execution, to avoid leaks.
Also, check for the validity of the session data member
in the observer methods, to make sure not to perform any
operations in case the session has been detached from
the observer.

Fixes: QTBUG-103239
Change-Id: I54cfd27b2971ec208e47f1b4038c526417c41a34
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit ac24ced62c4cfe83067765f3e049543319b092cb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>