Enable error reporting when failed to initialize backend element
This patch adds an option to report error message when a backend
component cannot be instantiated. This error message is then
displayed in the console with qWarning and in some cases reported
to the app user with error signals. We need further improvements
on the error reporting side.

Additionally, this patch cleans up the code in the .cpp API
classes.

Change-Id: Id39865cc8f1e9b52804bf5b9d9b15e738508f860
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit f1aa625049a08519d52ba87958e25cb80e47fd1e)