Simplify QMaybe and add specialization for pointer type
Simplifying QMaybe to only provide either value or error makes it more
readable and avoids problems for pointer types.

The patch add specialization for the pointer type. This type does not
require std::optional as it is optional in it self. Furthemore, using
the generic implementation may lead to errors, since the operator bool()
returns true for objects created with QMaybe(nullptr) which is counter
intuitive.

Change-Id: Idd03f7d1ae52abccc7c5b2298edc159899f13dd6
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>