Always create valid Codec and improve error reporting
Instead of relying on isValid() it is better to always create valid
Codec and if it cannot be created return error to the user. This way
we do not have to valide access to the d pointer in all member access
functions that was missing.

Fix AVCodecContext leak on some return paths from the Codec constructor.

Pick-to: 6.4
Change-Id: I6e8fa548e12194d826ae3198a84b066fdffe82d9
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>