Simplify logic for selecting streams when loading media
This patch simplifies logic for selecting streams for playback by
extracting it to a separate function getDefaultStreamIndex().
Additionally, it removes double booking of selected streams on
m_currentAVStreamIndex and replaces it with function computing the
requested indexes.

As a bonus, the patch fixes memleak that occurs when
avformat_find_stream_info fails by releasing context.

Pick-to: 6.4
Change-Id: I6a5073f160163082f01f67d1be5d6780955321de
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>