Android: Fix for setting video playback rate
According to doc[0], setting playback params may change the state of the
player. Calling it on a prepared MediaPlayer with non-zero speed is
equivalent to calling start(). This behavior is not expected in the Qt
implementation. Therefore, changing the input speed should only be done
in the Playing state or instead of calling play().

In such case, QtAndroidMediaPlayer need to handle state change,
notification and start progress watcher. Therefore, setting playback
params has been moved to QtAndroidMediaPlayer.

Also playbackRate() should not rely on value from Android MediaPlayer.
The value can be wrong for example when Video is in Pause state.

[0]https://developer.android.com/reference/android/media/MediaPlayer#setPlaybackParams(android.media.PlaybackParams)

Fixes: QTBUG-104670
Pick-to: 6.4 6.3 6.2
Change-Id: I5ce7696cf986a1e1e2aaa0035df806e172a5b7b7
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>