Look up camera format in IMFSourceReader formats
Unfortunately, the index of camera format in the
QCameraDevice::videoFormats() list may not match the index in
the IMFSourceReader list as some of the IMFSourceReader formats may
not be on the QCameraDevice list. Thus, we have to look up the Qt format
on the IMFSourceReader list. No shortcuts.

Change-Id: If99dc1b451a446a303bf516d4212a4eaf64bb52a
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
(cherry picked from commit 953468d895413c738e7c8784c271f394e6c53a8d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>