VideoOutput: Always update geometry when video surface format changes
When the video surface format changes, the video output's
_q_updateNativeSize() slot is called as a response. Inside this slot,
we mark the geometry as dirty (needing update) only if the native
size was changed.

However, this is not the only case that would require a geometry
update. In the case where the video surface format's scanline direction
changes, and this can happen when the pixel format changes (eg: for
uncompressed RGB formats the scanline direction depends on the sign of
the bitmap height, for YUV is always TopToBottom, etc.), the geometry
needs to update too, otherwise the video may appear vertically flipped.

To fix, always mark the geometry as dirty when inside the
_q_updateNativeSize() slot. As this is called only whenever the video
source format changes or the dimensions of the video output change,
it won't lead to many unnecessary calls to update the geometry.

Fixes: QTBUG-96746
Change-Id: I7d1d3a0b8e7b7122b7e64dfdd7c1de4098d7daf8
Reviewed-by: Lars Knoll <lars@knoll.priv.no>