QAVFCameraBase: Get max digital zoom factor from the active format
The camera device's maxAvailableVideoZoomFactor is equal to the
device's format videoMaxZoomFactor value on single camera devices, but
this is not the case for devices with dual/triple cameras. If we then
set the device's zoom factor to a value bigger than the format's
videoMaxZoomFactor, this can lead to a crash. So to be on the safe
side, get the maximum zoom factor from the format rather than device.

Also, directly set the videoZoomFactor of the device if the rate is 0,
since setting the factor using rampToVideoZoomFactor:withRate: with a
rate of 0, doesn't have any effect.

Fixes: QTBUG-108009
Change-Id: Icf7fcdacefd3b1e0084e8ef6cf4de97c6b1b4996
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit bc01f4e7c459a591f476dbd23d95ead4c48035ea)
Reviewed-by: Doris Verria <doris.verria@qt.io>