Allow sinks that do not have associated QRhi to receive video frames
AndroidSurfaceTexture was created only when QVideoSink has an associated
QRhi. This enables QVideoSink to display textures created by camera or
media player as the textures are created in the OpenGL context
associated with QRhi. However, we can allow QVideoSink without QRhi to
receive frames as the context sharing can be established later or
frames are processed as QImage.

Fixes: QTBUG-108027
Pick-to: 6.4
Change-Id: Iac8925da43aa100fe03c09f57f3d254c27add49c
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>