Remove commented API for setting set of QVideoSink as output for player
Reintroducing this API is non-trivial as the current video sink is
the main surface that media player and camera view finder render to and
depend on. Media player sets up its rendering pipeline to use
the rendering hardware associated with the video sink. Thus, when using
multiple sinks, one must be selected. Hence, this API does not match
well the underlying system. Displaying video in multiple video sinks is
easy to achieve with the technique described in the related QTBUG.

Fixes: QTBUG-95127
Pick-to: 6.4 6.3 6.2
Change-Id: I751f9cac4a02d9d97c5c386f814cc9391ae73c69
Reviewed-by: Lars Knoll <lars@knoll.priv.no>