Fix ffmpeg camera crash on macOS (double releasing of an object)
The reason for the crash is 'autorelease' and 'release' in the
destructor. In applications, QMacAutoReleasePool releases it 2nd time.
The application crashed after a few camera changes.
So we should use only one approach. The suggestion is to manage
the object manually for both backends in order to make the behavior
more predictable.
The added test checks the case.

Pick-to: 6.4
Change-Id: I80a644acd94ae469a16fd95ba971441c78e7a700
Reviewed-by: Doris Verria <doris.verria@qt.io>