Fix crash on macos13 with iphone camera
On macos13 new feature has been introduced:
possibility using a camera of paired iphone on macos.

The feature works but it's a bit buggy: AVCaptureDevice.hasFlash is true
but none of flash modes is supported.

Setting of not supported flash mode causes exception.
The exeptions happens when setting AVCaptureFlashModeOff, but,
in theory, it can occur in other corner cases.
Documentation recommentds to check if flash mode supported in order to
avoid exceptions.

So, what's done:
- checking of supported flash and torch modes before setting
- use captureDevice.isFlashAvailable not only for mac os since
  ios camera can be paired

Task-number: QTBUG-108018
Change-Id: I42772edb4c26481283d1bd321914bf40284efde9
Reviewed-by: Doris Verria <doris.verria@qt.io>