Correctly finish, clear and delete AndroidTextureThread
'QAndroidTextureVideoOutput' destructor waited for 'm_surfaceThread' to
be finished but seems that this thread was nowhere requested to finish.
In addition, 'AndroidTextureThread' created 'std::unique_ptr<QRhi> m_rhi;'
in it's own thread but didn't reset it. Thus, deleting the whole
'AndroidTextureThread' from another thread, started deleting it and
failed.

Task-number: QTBUG-109168
Pick-to: 6.4 6.5
Change-Id: I9e12228c579457710c4fe5d8185bffabf9f3755a
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>