Android: fix thread hang on tst_qvideowidget
QVideoWidget gets blocked because of a hanging thread on
QAndroidTextureVideoOutput. This thread waits on the destructor forever.
This only happened when the provided sink had no RHI.
QAndroidTextureVideoOutput depends on OpenGL so it does not work without
an RHI. Therefore a check was added on the constructor to prevent the
thread to start in this case.

Fixes: QTBUG-109443
Change-Id: I158ff95967116697018e7d6f34c29c5e999182be
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit f6b0e3a33add68972e3ba556dddb1f745224ce01)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>