Fix memleak in QGstCaps
gst_device_get_caps and gst_pad_query_caps require decreasing of
reference count when caps are not longer needed. This was not done
with QGstCaps. This patch adds reference counting on QGstCaps and
remove QGstCapsMutable as it was based on a missconception. GstCaps can
be mutated only when refernce count is 1. This was not ensured with
QGstCapsMutable.

Pick-to: 6.5 6.4 6.2
Change-Id: Id058a0a916b5f76eef758a099791bec7cc917af9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>