Implement utility for checking and dumping audio config on macOS
What is the utility able to do:
- Enquires a vast set of audio devices configurations via
AudioObjectGetPropertyData.
- dumps results.
- check the configuration many times in order to check if it's stable.

Why do we need it:
  We have bugs with QtMM on macOS with audio input/output that
are not reproducible with development environments.
Pretty often such bugs have smth to do with specific audio devices on
the bug reporter's side, and it's unclear what AudioObjectGetPropertyData
returns in these cases.
  When I was investigating previous bugs, I attached such a code to
the ticket so that bug reporters could provide more info.
  In light of the above, it's reasonable to have such a utility in
the qt repo. Having it now in a single file could simplify its
compilation and run by customers (just one command in README).

Pick-to: 6.5 6.4
Task-number: QTBUG-108176
Change-Id: I3560353c78a0852a426f25b3487ab0a93e39c03f
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>