Android: Prevent Camera set automatically active
The Android backend was different from the rest of the backends and it
automatically started the camera when it was attached to media capture
session. This patch changes to set it as the rest of the backends.

Fixes: QTBUG-109415
Change-Id: I88342e9db068bec05404f5bfd91f73ea5015bb96
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit f18be9c59cb38fb3cc5046932c5587997af35cf5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>