Fix crash with ffmpeg on specific video files
FFmpeg doc says that the following case is possible:
AVFrame.linesize[i] < 0 and AVFrame.data[i] points to the end of data,
it's a weird case of flipped pixels direction, only one video with
such an encoding has been found for now.

Function sws_scale solves the problem so we should call it
in the cases of flipped videos.

I'm going to add a test after finding a way of getting such files.

Task-number: QTBUG-108403
Pick-to: 6.5 6.4
Change-Id: I72f45781e2c3983d6a1991db2b833b78184b90dd
Reviewed-by: Lars Knoll <lars@knoll.priv.no>