Android: fix for randomly not working camera
After commit: 2abfa1e5cd3cbb148305dc0fe85671488a542397, there was a
thread race in QAndroidCameraSession. Calling setVideoSize put
setFrameSize method to queue (which resets a viewport). It happened from
time to time that source texture coordinates was calculated with not
ready viewport.

To avoid this situation setFrameSize will be queued as
BlockingQueuedConnection.

Pick-to: 6.5 6.4
Fixes: QTBUG-109561
Fixes: QTBUG-109391
Change-Id: Ie6a7bbd332fb42f1aba4fea53449ecac082aaf89
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>