Android: Remove deadlock from QAndroidCamera
During the destruction of QAndroidCamera, stopBackgroundThread was
called with blocked rwLock (QReadWriteLock). stopBackgroundThread is
waiting on join() for mBackgroundThread. There is a big possibility for
mBackgroundThread to stuck on rwLock (e.g. while handling
onFrameAvailable). Because of that, mBackgroundThread cannot be joined.
This leads to a deadlock.

This commit remove the deadlock problem by releasing QReadWriteLock
before stopBackgroundThread method is called.

Fixes: QTBUG-110285
Change-Id: Ie0e4ed26d5bb2c2408460d3884430a8a659c4d10
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>