Android: don't change Camera State for each frame
onCaptureStarted, onCaptureCompleted and onCaptureFailed methods are
called for each frame. Camera state (camera activity relies on it)
should not be updated with such frequency.

Remove onCaptureSessionCompleted and onCaptureSessionStarted handling,
as the only purpose of them were to set a camera state.

Remove camera status change for onCaptureFailed, as other requests are
unaffected with this failure

Instead of CaptureCallback start to use StateCallback for Camera state
updates (onActive and onClosed methods)

Fixes: QTBUG-110317
Change-Id: I9af1399033e24dbbf395f55ae23e1696aa4d42ce
Reviewed-by: Samuel Mira <samuel.mira@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>