Fix memory leak and random crash in windows playback backend
The leak is trivial, just not deleted object.

The crash is not trivial for reproducing; the fix is based on
the assumption that some events can be enqueued for an object
between deleteLater and real deleting.

Task-number: QTBUG-109149
Change-Id: I256f29a03e7f01d27923b183b63d9a6b6a4f21b9
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit de266fd951d56dd937d63e79c6adf1702fbe0ec9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>